GIX-1555 Pass confirmation_text on notifyParticipation (#2558)
# Motivation

SNS creators can specify a confirmation_text that needs to be confirmed
by participants.
This text needs to be included in the notifyParticipation call so the
backend can verify the confirmation regardless of which client the
participant uses.

# Changes

1. Make the `confirmationText` parameter in the `notifyParticipation`
api non-optional.
2. Get the confirmation text from the project summary and pass it in the
`notifyParticipation` call.
3. In the api test, pass `confirmationText` and verify that it is passed
on to the canister.
4. In `frontend/src/tests/lib/services/sns-sale.services.spec.ts` verify
that `confirmationText` is passed on whether it's set or empty and with
both initial participation and restored participation.
5. Also remove `snsQueryStore.setData(` which was unused because the
project store is mocked.

# Tests

Unit test + manual testing.