Fix reproducibility by not explicitly importing cjs modules from agent-js in esm (#2584)
# Motivation

@dskloetd figured out that our frontend reproducibility issue started
with PR #2377. After having a look at the PR we noticed it introduced an
explicit import of a CJS module of agent-js within the code that is
bundled with ESM. As I suspected that our issue was linked to pollyfying
the Node modules for the browser, we replaced the import by inline code
and as a result, the reproducibility was improved.

Using the `./script/test-reproducibility` script, on my machine, I am
now able to run 100 successful check in a row.

# Changes

- remove explicit CJS `import`
- use SvelteKit `browser` information when possible
- when not (see PR for explanation) use a custom `isBrowser` information