E2E: Use test step to follow progress when headless (#2590)
# Motivation

If a test uses `test.step` you can see the name of the step in the
terminal when running a headless e2e test.
The intended way to do this is to put all the code for a step inside a
closure passed to `test.step()` but this makes the code hard to read and
prevents sharing constants created in one step to another step.

# Changes

1. Add a utility function that lets you call `test.step` without passing
a closure. It simulates that all the code until the next call to
`step()` runs inside the passed closure.
2. Use `step()` inside existing e2e tests to make it easy to follow
progress.

# Tests

Tested by running `npm run test-e2e` and observing the result in the
terminal.

<img width="1263" alt="image"
src="https://github.com/dfinity/nns-dapp/assets/122978264/7263ebac-22dd-4d9d-ae5f-ff6e3cbb4801">