Commit in metadata (#2500)
# Motivation
Requests from the nns & testing teams to publish the commit of our
canisters. In fact, it is becoming common to publish both the commit and
the API did file so we cover both on both canisters. Four with one
stroke.

# Changes
* Created a new command to add metadata. This is NOT in the canister
build command so that the canister build can be cached, and commits that
do not affect a canister wasm don't cause a rebuild just because the
commit has changed.
* Created a test for the command
* Called the command in the dockerfile
* Created a test for deployments from the docker build

# Tests
* Tests are included.  See CI for the results.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>