feat: display a warning after sign-in for testing in mainnet (#2570)
# Motivation

We aim to deploy on mainnet a version of NNS-dapp that is only meant for
testing purpose. That's why we want to display a warning after sign-in.

# Notes

PR provide the UI feature. The flag to activate the function is here
disabled.

I was first thinking at a feature flag but, actually we probably gonna
display it only for a particular environment. Therefore it shouldn't be
a flag but dependending of the `DFX_NETWORK` that will be set up once
the version will really be deployed.

# Changes

- new warning component that renders a modal after sign-in for the app
pages (not login page)

# Screenshot

<img width="1536" alt="Capture d’écran 2023-05-24 à 15 16 37"
src="https://github.com/dfinity/nns-dapp/assets/16886711/b06f3ab4-60f2-4f9d-89ca-17e876e4e0f6">

<img width="1536" alt="Capture d’écran 2023-05-24 à 15 16 24"
src="https://github.com/dfinity/nns-dapp/assets/16886711/521c279b-788c-4777-82e3-3ca638621908">