Expose ledger canister id in sns summary (#2642)
# Motivation

Regardless if Sns, ckBTC or ICP we want to be able to fetch balance and
transactions uses the ICRC-1 ledger.

In Snses we generally use the "wrapper" to access the canister but, in
the particular crontab feature #2639 we want to avoid the use of such
object and use `@dfinity/ledger` library directly regardless of the
context.

That is why we need to access the ledger canister Id of a Sns project.

# Changes

- export `ledgerCanisterId` in `SnsSummary` (as we already did for swap
and governance)