Apple install to home screen look and feel (#2645)
# Motivation

The new release has been deployed on the mainnet, allowing me to test
the installation of the dapp on the iOS home screen. However, it seems
that Apple displays the icons over a black background (ignoring theming
properties). Since the icon is designed as a circle, it doesn't look
visually appealing.

# Screenshot


![IMG_0534](https://github.com/dfinity/nns-dapp/assets/16886711/4d6edc28-8c7f-45f6-8e30-8b3352aa8d69)