Call ckBTC update balance in withdrawal account process (#2646)
# Motivation

Calling ckBTC `updateBalance` has nothing to do with the withdrawal
account process but, because users are confused about when and how to
call it, product team required the addition of this additional no
related call within the process. That way, when user navigate once per
session to the ckBTC accounts page, the update balance call is also
triggered.

# Changes

- call `updateBalance` within component `CkBTCWithdrawalAccount`
- add option `uiIndicators` to `updateBalance` because we do not want to
display a spinner or success toast in that particular case