Use ckBTC info for retrieve min amount (#2681)
# Motivation

Use new ckBTC info fetched from minter for the minimum amount that can
be converted check instead of a constant.

# Note

I am not super super happy about displaying an error when the minimal
amount information have not been certified yet but, I've tried several
implementation and UX and it ultimately felt like the easiest and
cleanest way to handle what is, at the end of the day, an edge case.

# PRs

- [x] #2682

# Changes

- remove constant `RETRIEVE_BTC_MIN_AMOUNT` and replace with new
`MinterInfo` data
- enhance `assertCkBTCUserInputAmount` to use the data from the store
and checks if these are provided and certified