GIX-1584: Build env var being used in worker (#2684)
# Motivation

As described in https://dfinity.atlassian.net/browse/GIX-1584 a build
environment variables was being used instead of the one from the
initialization.

The problem was that the workers can't access `window` and therefore do
not have access to the initialization env vars.

Solution for now: Pass the parameter in the post message instead of
accessing the env vars directly.

Food for thought: Should we substitute the variables in the JS files
instead of setting them in HTML during canister installation and
upgrade?

# Changes

* `queryCanisterDetails` expects two new optional params: `host` and
`fetchRootKey`.
* PostMessage to the cycles worker includes `host` and `fetchRootKey`.
* Cycles service `initCyclesWorker` passes `host` and `fetchRootKey` in
the post message to the worker.
* Cycles worker uses new params `host` and `fetchRootKey` when calling
`queryCanisterDetails`.

# Tests

* New test case in `queryCanisterDetails` to check that new param is
being used.