Controller may upload tarball of assets (#2632)
# Motivation
We wish to be able to deploy the nns-dapp to an application subnet.
This, however, means that it needs to be deployed in parts. Deploying
parts requires authorization. A controller has full control over the
canister anyway, so this is an easy case.

- References: https://dfinity.atlassian.net/browse/GIX-1588

# Changes
- Add an API method that allows the controller of a canister to upload
an assets tarball.

# Warning
- [x] Before this can be used in CI, we need to bump the dfx version to
one that uses a newer replica. The current dfx uses a very old replica.
  - [x] Bump dfx to 0.14.1 to support is_controller
  - [x] Bump snsdemo dfx to 0.14.1
- [x] Likewise it would be prudent to confirm the replica version used
on the NNS subnet.

# Tests
A test is included but it cannot be run on CI, as CI doesn't support
`is_controller()`. The testnets do support it. Application subnets vary
in which replica commit they use; theoretically we could create a
canister on a very old subnet and then be unable to install assets via
this path until the replica has been upgraded. I have deployed the
nns-dapp to a static testnet, though, and run the test against that
deployment.

Merging this without a feature flag requires a dfx version with this
commit in the replica:
```
commit c8f3c400c10adc0a44dadcc0dcffbf812a37ce32
Merge: 76ccbe7f90 8f4a04d799
Author: Dragoljub Duric <dragoljub.duric@dfinity.org>
Date:   Thu Mar 30 13:27:59 2023 +0000

    Merge branch 'EXC-1351-implement-is-controller-system-api' into 'master'
    
    feat: Implement is_controller system api
    
    Closes EXC-1351.
    [Description](https://github.com/dfinity/interface-spec/pull/141).
    
    Closes EXC-1351
    
    See merge request dfinity-lab/public/ic!11411
```

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>