Render proposal arguments with more than one field (#2716)
# Motivation
There was a design error, in which I was under the impression that
proposal arguments were represented by IDLTypes. I knew that proposals
could have multiple arguments but was under the impression that a vector
of values was itself an IDLType. For a long time this fallacy seemed to
be correct as proposals most commonly have one argument. We recently
encountered a proposal with no arguments and this broke the proposal
rendering. On investigation, there is a type called IDLArgs that is
roughly equivalent to a vector of IDLTypes but that is not itself an
IDLType. It is different from an IDLType vector in that in a vector, the
types of the contents are all the same whereas for IDLArgs, they can be
different. In the common case, where there is a single type, the two
behave in much the same way. This is how the false assumption passed
scrutiny for so long.

A quick fix was added that handles the case where there are no arguments
and tests were added for the cases where there are multiple arguments.
However in the case of multiple arguments, only the first was rendered.

The idl2json library how offers a proper solution, in which `IDLArgs`
are parsed properly, so we can use that and remove the hack. This does
however change the output. Previously the JSON output was a single
JSONValue corresponding to the first argument. Now the output is an
array of JSONValues, with as many entries as there are arguments.

# Changes
- Update the idl2json library version.
- Replace the partial logic for handing variable args with the library
call that handles them all correctly.

# Questions
- Does anything need to be updated with the front-end proposal
rendering?

# Tests
There are existing unit tests for the conversion. These have been
updated.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>