Fix: Navigate from SNS proposal to launchpad (#2723)
# Motivation

There was a bug when the user navigated from the SNS Proposal detail
page to the launchpad. The user was redirected to the proposals page.

The issue is a problem with Svelte transitions in components that
persist across routes. In our case the `TotalValueLocked.svelte`.

With this Svelte issue, there might be two pages rendered at the same
time. Therefore, when the user clicks in the Launchpad the
Launchpad.svelte is rendered, but the PropoposalDetail.svelte has not
yet been unmounted.

In this case, the ProposalDetail.svelte was rerendered with NNS
universe, and the sns proposal id. That triggered to render the
NnsProposalDetail page with the SNS proposal id. That means that the
NnsProposalDetail was trying to find the proposal with id of the SNS
proposal id. Yet, NNS removes old proposal, and the call to get the nns
proposal was returning not found. That triggered the error flow or
redirecting the user back to the Proposals page.

The fix was to use CSS transition instead of a Svelte transition in that
component.

# Changes

* Remove Svelte transition in `TotalValueLocked` and use a CSS
transition instead.

# Tests

* Fix tests after refactor.