TVL in various currencies (#2695)
# Motivation

It was requested to display the TVL in various currencies, so this PR
introduces these with a random query (since the TVL is fetched once per
session).

# Note

This PR does not modify the particular i18n of the display, we keep the
custom display of the TVL but updates the symbol according currency.

# Changes

- update did and canister for up-to-date definition of the TVL canister
- implement new option `currency`
- add `Math.random` to worker in order to get once per session a
different currency

# Screenshot


![gif](https://github.com/dfinity/nns-dapp/assets/16886711/d2452ede-f05a-49cf-9486-9d9f71eceafb)