Cmc url (#2737)
# Motivation
We use the NNS cycle minting and ledger canisters in the nns-dapp repo,
however these canisters are not listed in dfx.json. This means that the
candid files are not available to make API calls. It also means that the
canister IDs are not obtained in the standard way, not that this really
matters much but it would be a bit surprising for someone dipping into
the codebase.

# Changes
- Add entries for the NNS leger and cycles minting canisters in
dfx.json.
- Obtain their canister IDs using the usual canister remote ID logic.

# Tests
- The config should still have the correct canister IDs. There is an
existing test for this.

---------

Co-authored-by: Max Murphy-Skvorzov <max@winning.black>