GIX-1637: Show Vesting Period in SNS Neuron (#2743)
# Motivation

SNS neurons for developers might have a vesting period. During this
vesting period, the functionality of the neuron is limited.

Therefore, the UI should tell the user how long is until the neuron will
have it's full capability.

The UI might change. Product is looking into it but it shouldn't be very
different. We can change after releasing the feature if needed.

# Changes

* New SnsNeuronVestingPeriodRemaining to render the vesting period if
still active.
* Use the new component in the SnsNeuronMetaInfoCard.
* New sns neuron utils `isVesting` and `vestingInSeconds`.

# Tests

* Test the new component SnsNeuronVestingPeriodRemaining
* Add a test case in SnsNeuronMetaInfoCard
* Test new sns neuron utils.