Fix args verification in release template (#2747)
# Motivation

https://github.com/dfinity/nns-dapp/pull/2236 intended to add arguments
verification instructions to the proposal template but it actually only
added them to the rollback proposal template.

Also the instructions there suggest to check the sha256sum but the
proposal payload contains the binary hex so I think it makes more sense
to provide instructions to output that because that's what can actually
be compared to the proposal payload, as displayed on the dashboard, for
example.

# Changes

1. Change the instructions to output binary hex.
2. Copy the same instructions from `ROLLBACK.md` to `PROPOSAL.md`
3. For convenience, mention the `arg_hex` field in the proposal payload
to compare to.

# Tests

Ran `scripts/nns-dapp/release-template` and checked the sections in
`PROPOSAL.md`:
````
You may also want to verify the canister arguments.  In the proposal they are
binary, which is not very readable. Docker provides both binary and text formats
and you can verify that the text format corresponds to the binary `arg_hex`
field in the proposal.

```
cat nns-dapp-arg-mainnet.did
didc encode "$(cat nns-dapp-arg-mainnet.did)"
```
````
and `ROLLBACK.md`:
````
You may also want to verify the canister arguments.  In the proposal they are
binary, which is not very readable. Docker provides both binary and text formats
and you can verify that the text format corresponds to the binary `arg_hex`
field in the proposal.

```
cat nns-dapp-arg-mainnet.did
didc encode "$(cat nns-dapp-arg-mainnet.did)"
```